% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miles_to_km.R
\name{miles_to_km}
\alias{miles_to_km}
\title{Miles to km}
\usage{
miles_to_km(x)
}
\arguments{
\item{x}{Numeric in miles}
}
\value{
A value in km
}
\description{
This function converts the values from miles to kilometers.
}
\details{
If you use this and argument, this and that happens...
}
\examples{
miles_to_km(2)
miles_to_km(0)
miles_to_km(100)
}
\references{
Smith, J., 2024, Afdsfdsfs, ADfsds, doi:
}
