% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{temp_conversion}
\alias{temp_conversion}
\title{Temperature Conversion}
\usage{
temp_conversion(temp_f)
}
\arguments{
\item{temp_f}{a vector containing temperature values
in Fahrenheit}
}
\value{
a numeric vector
}
\description{
This function converts temperature
from Fahrenheit to Celsius.
}
\examples{
temp_conversion(75)
temp_conversion(110)
temp_conversion(0)
temp_conversion(c(0, 75, 110))
}
